+++ Assembly/Whiskey/RequiresNuGetPackageAttribute.cs
@@ -1,0 +1,10 @@
+
+namespace Whiskey
+{
+    public sealed class RequiresNuGetPackageAttribute : RequiresToolAttribute
+    {
+        public RequiresNuGetPackageAttribute(string name) : base(name, "NuGet")
+        {
+        }
+    }
+}