+++ Assembly/Whiskey/RequiresPowerShellModuleAttribute.cs
@@ -3,7 +3,7 @@ namespace Whiskey
 {
     public sealed class RequiresPowerShellModuleAttribute : RequiresToolAttribute
     {
-        public RequiresPowerShellModuleAttribute(string moduleName) : base(moduleName)
+        public RequiresPowerShellModuleAttribute(string moduleName) : base(moduleName, "PSModule")
         {
         }