+++ Whiskey/Functions/Install-WhiskeyTool.ps1
@@ -6,46 +6,61 @@ function Install-WhiskeyTool
     Downloads and installs tools needed by the Whiskey module.
 
     .DESCRIPTION
-    The `Install-WhiskeyTool` function downloads and installs PowerShell modules or NuGet Packages needed by functions in the Whiskey module. PowerShell modules are installed to a `Modules` directory in your build root. If PowerShell modules are already installed globally and are listed in the PSModulePath environment variable, they will not be re-installed. A `DirectoryInfo` object for the downloaded tool's directory is returned.
+    The `Install-WhiskeyTool` function downloads and installs PowerShell modules or NuGet Packages needed by functions
+    in the Whiskey module. PowerShell modules are installed to a `Modules` directory in your build root. If PowerShell
+    modules are already installed globally and are listed in the PSModulePath environment variable, they will not be
+    re-installed. A `DirectoryInfo` object for the downloaded tool's directory is returned.
 
-    `Install-WhiskeyTool` also installs tools that are needed by tasks. Tasks define the tools they need with a [Whiskey.RequiresTool()] attribute in the tasks function. Supported tools are 'Node', 'NodeModule', and 'DotNet'.
+    `Install-WhiskeyTool` also installs tools that are needed by tasks. Tasks define the tools they need with a 
+    [Whiskey.RequiresTool()] attribute in the tasks function. Supported tools are 'Node', 'NodeModule', and 'DotNet'.
 
-    Users of the `Whiskey` API typcially won't need to use this function. It is called by other `Whiskey` function so they have the tools they need.
+    Users of the `Whiskey` API typcially won't need to use this function. It is called by other `Whiskey` function so
+    they have the tools they need.
 
     .EXAMPLE
     Install-WhiskeyTool -NugetPackageName 'NUnit.Runners' -version '2.6.4'
 
-    Demonstrates how to install a specific version of a NuGet Package. In this case, NUnit Runners version 2.6.4 would be installed.
+    Demonstrates how to install a specific version of a NuGet Package. In this case, NUnit Runners version 2.6.4 would
+    be installed.
     #>
     [CmdletBinding()]
     param(
         # The attribute that defines what tool is necessary.
-        [Parameter(Mandatory, ParameterSetName='Tool')]
+        [Parameter(Mandatory, ParameterSetName='FromAttribute')]
         [Whiskey.RequiresToolAttribute] $ToolInfo,
 
-        # The directory where you want the tools installed.
-        [Parameter(Mandatory, ParameterSetName='Tool')]
-        [String] $InstallRoot,
-
         # The task parameters for the currently running task.
-        [Parameter(Mandatory, ParameterSetName='Tool')]
+        [Parameter(Mandatory, ParameterSetName='FromAttribute')]
         [hashtable] $TaskParameter,
 
         # Running in clean mode, so don't install the tool if it isn't installed.
-        [Parameter(ParameterSetName='Tool')]
+        [Parameter(ParameterSetName='FromAttribute')]
         [switch] $InCleanMode,
 
         # The path to a directory where downloaded package files should be saved prior to installation.
-        [Parameter(Mandatory, ParameterSetName='Tool')]
+        [Parameter(Mandatory, ParameterSetName='FromAttribute')]
+        [Parameter(ParameterSetName='AtRuntime')]
         [String] $OutFileRootPath,
+
+        [Parameter(ParameterSetName='AtRuntime')]
+        [AllowEmptyString()]
+        [String] $ProviderName,
+
+        [Parameter(Mandatory, ParameterSetName='AtRuntime')]
+        [String] $Name,
 
         # The name of the NuGet package to download.
         [Parameter(Mandatory, ParameterSetName='NuGet')]
         [String] $NuGetPackageName,
 
-        # The version of the package to download. Must be a three part number, i.e. it must have a MAJOR, MINOR, and BUILD number.
         [Parameter(ParameterSetName='NuGet')]
+        [Parameter(ParameterSetName='AtRuntime')]
         [String] $Version,
+
+        # The directory where you want the tools installed.
+        [Parameter(Mandatory, ParameterSetName='FromAttribute')]
+        [Parameter(Mandatory, ParameterSetName='AtRuntime')]
+        [String] $InstallRoot,
 
         # The root directory where the tools should be downloaded. The default is your build root.
         #
@@ -93,114 +108,110 @@ function Install-WhiskeyTool
 
         if( $PSCmdlet.ParameterSetName -eq 'NuGet' )
         {
-            try
-            {
-                if( -not $IsWindows )
-                {
-                    Write-WhiskeyError -Message ('Unable to install NuGet-based package {0} {1}: NuGet.exe is only supported on Windows.' -f $NuGetPackageName,$Version) -ErrorAction Stop
-                    return
-                }
-                $packagesRoot = Join-Path -Path $DownloadRoot -ChildPath 'packages'
-                $version = Resolve-WhiskeyNuGetPackageVersion -NuGetPackageName $NuGetPackageName -Version $Version -NugetPath $whiskeyNuGetExePath
-                if( -not $Version )
-                {
-                    return
-                }
-
-                $nuGetRootName = '{0}.{1}' -f $NuGetPackageName,$Version
-                $nuGetRoot = Join-Path -Path $packagesRoot -ChildPath $nuGetRootName
-                Set-Item -Path 'env:EnableNuGetPackageRestore' -Value 'true'
-                if( -not (Test-Path -Path $nuGetRoot -PathType Container) )
-                {
-                & $whiskeyNuGetExePath install $NuGetPackageName -version $Version -outputdirectory $packagesRoot | Write-CommandOutput -Description ('nuget.exe install')
-                }
-                return $nuGetRoot
-            }
-            finally
-            {
-                if( (Test-Path 'env:APPVEYOR') )
-                {
-                    $Global:Error | Format-List * -Force | Write-Verbose -Verbose
-                }
-            }
+            $msg = 'The "Install-WhiskeyTool" function''s "NuGetPackage" name parameter is obsolete. Use ' +
+                   '[Whiskey.WhiskeyTool] attribute on your task instead.'
+            Write-Warning -Message $msg
+                
+            Install-WhiskeyNuGetPackage -Name $NuGetPackageName -Version $Version -BuildRootPath $DownloadRoot
+            return
         }
-        elseif( $PSCmdlet.ParameterSetName -eq 'Tool' )
+
+        if( $PSCmdlet.ParameterSetName -eq 'FromAttribute' )
         {
-            $provider,$name = $ToolInfo.Name -split '::'
-            if( -not $name )
+            $ProviderName = $ToolInfo.ProviderName
+            $Name = $ToolInfo.Name
+            $Version = $TaskParameter[$ToolInfo.VersionParameterName]
+            if( -not $Version )
             {
-                $name = $provider
-                $provider = ''
-            }
-
-            $version = $TaskParameter[$ToolInfo.VersionParameterName]
-            if( -not $version )
-            {
-                $version = $ToolInfo.Version
-            }
-
-            if( $ToolInfo -is [Whiskey.RequiresPowerShellModuleAttribute] )
-            {
-                $module = Install-WhiskeyPowerShellModule -Name $name `
-                                                          -Version $version `
-                                                          -BuildRoot $InstallRoot `
-                                                          -SkipImport:$ToolInfo.SkipImport `
-                                                          -ErrorAction Stop
-
-                if( $ToolInfo.ModuleInfoParameterName )
-                {
-                    $TaskParameter[$ToolInfo.ModuleInfoParameterName] = $module
-                }
-
-                return
-            }
-
-            $toolPath = $null
-            switch( $provider )
-            {
-                'NodeModule'
-                {
-                    $nodePath = Resolve-WhiskeyNodePath -BuildRootPath $InstallRoot
-                    if( -not $nodePath )
-                    {
-                        Write-WhiskeyError -Message ('It looks like Node isn''t installed in your repository. Whiskey usually installs Node for you into a .node directory. If this directory doesn''t exist, this is most likely a task authoring error and the author of your task needs to add a `WhiskeyTool` attribute declaring it has a dependency on Node. If the .node directory exists, the Node package is most likely corrupt. Please delete it and re-run your build.') -ErrorAction stop
-                        return
-                    }
-                    $toolPath = Install-WhiskeyNodeModule -Name $name `
-                                                          -BuildRootPath $InstallRoot `
-                                                          -Version $version `
-                                                          -Global `
-                                                          -InCleanMode:$InCleanMode `
-                                                          -ErrorAction Stop
-                }
-                default
-                {
-                    switch( $name )
-                    {
-                        'Node'
-                        {
-                            $toolPath = Install-WhiskeyNode -InstallRootPath $InstallRoot `
-                                                            -Version $version `
-                                                            -InCleanMode:$InCleanMode `
-                                                            -OutFileRootPath $OutFileRootPath
-                        }
-                        'DotNet'
-                        {
-                            $toolPath = Install-WhiskeyDotNetTool -InstallRoot $InstallRoot -WorkingDirectory (Get-Location).ProviderPath -Version $version -ErrorAction Stop
-                        }
-                        default
-                        {
-                            throw ('Unknown tool "{0}". The only supported tools are "Node" and "DotNet".' -f $name)
-                        }
-                    }
-                }
-            }
-
-            if( $ToolInfo.PathParameterName )
-            {
-                $TaskParameter[$ToolInfo.PathParameterName] = $toolPath
+                $Version = $ToolInfo.Version
             }
         }
+
+        if( -not $OutFileRootPath )
+        {
+            $OutFileRootPath = Join-Path -Path $InstallRoot -ChildPath '.output'
+        }
+
+        if( -not (Test-Path -Path $OutFileRootPath) )
+        {
+            New-Item -Path $OutFileRootPath -ItemType 'Directory' | Out-Null
+        }
+
+        if( $ToolInfo -is [Whiskey.RequiresPowerShellModuleAttribute] )
+        {
+            $module = Install-WhiskeyPowerShellModule -Name $Name `
+                                                      -Version $Version `
+                                                      -BuildRoot $InstallRoot `
+                                                      -SkipImport:$ToolInfo.SkipImport `
+                                                      -ErrorAction Stop
+
+            if( $ToolInfo.ModuleInfoParameterName )
+            {
+                $TaskParameter[$ToolInfo.ModuleInfoParameterName] = $module
+            }
+
+            return
+        }
+
+        $toolPath = $null
+        switch( $ProviderName )
+        {
+            'NuGet'
+            {
+                $toolPath = Install-WhiskeyNuGetPackage -Name $Name -Version $Version -BuildRootPath $InstallRoot
+            }
+            'NodeModule'
+            {
+                $nodePath = Resolve-WhiskeyNodePath -BuildRootPath $InstallRoot
+                if( -not $nodePath )
+                {
+                    $msg = 'It looks like Node isn''t installed in your repository. Whiskey usually installs Node.js ' +
+                           'for you into a .node directory. If this directory doesn''t exist, this is most likely a ' +
+                           'task authoring error and the author of your task needs to add a `WhiskeyTool` attribute ' +
+                           'declaring it has a dependency on Node.js. If the .node directory exists, the Node ' +
+                           'package is most likely corrupt. Please delete it and re-run your build.'
+                    Write-WhiskeyError -Message $msg -ErrorAction Stop
+                    return
+                }
+                $toolPath = Install-WhiskeyNodeModule -Name $Name `
+                                                      -BuildRootPath $InstallRoot `
+                                                      -Version $Version `
+                                                      -Global `
+                                                      -InCleanMode:$InCleanMode `
+                                                      -ErrorAction Stop
+            }
+            default
+            {
+                switch( $Name )
+                {
+                    'Node'
+                    {
+                        $toolPath = Install-WhiskeyNode -InstallRootPath $InstallRoot `
+                                                        -Version $Version `
+                                                        -InCleanMode:$InCleanMode `
+                                                        -OutFileRootPath $OutFileRootPath
+                    }
+                    'DotNet'
+                    {
+                        $toolPath = Install-WhiskeyDotNetTool -InstallRoot $InstallRoot `
+                                                              -WorkingDirectory (Get-Location).ProviderPath `
+                                                              -Version $Version `
+                                                              -ErrorAction Stop
+                    }
+                    default
+                    {
+                        throw "Unknown tool ""$($Name)"". The only supported tools are ""Node"" and ""DotNet""."
+                    }
+                }
+            }
+        }
+
+        if( $PSCmdlet.ParameterSetName -eq 'FromAttribute' -and $ToolInfo.PathParameterName )
+        {
+            $TaskParameter[$ToolInfo.PathParameterName] = $toolPath
+        }
+
+        return $toolPath
     }
     finally
     {