+++ Whiskey/Tasks/NuGetRestore.ps1
@@ -2,24 +2,29 @@
 function Restore-WhiskeyNuGetPackage
 {
     [CmdletBinding()]
-    [Whiskey.TaskAttribute("NuGetRestore",Platform='Windows')]
+    [Whiskey.TaskAttribute('NuGetRestore', Platform='Windows')]
+    [Whiskey.RequiresNuGetPackage('NuGet.CommandLine', Version='6.*', PathParameterName='NuGetPath')]
     param(
         [Parameter(Mandatory)]
         [Whiskey.Tasks.ValidatePath(Mandatory)]
-        [String[]]$Path,
+        [String[]] $Path,
 
-        [String[]]$Argument,
+        [String[]] $Argument,
 
-        [String]$Version,
+        [String] $Version,
 
-        [Whiskey.Tasks.ParameterValueFromVariable('WHISKEY_BUILD_ROOT')]
-        [IO.DirectoryInfo]$BuildRoot
+        [String] $NuGetPath
     )
 
     Set-StrictMode -Version 'Latest'
     Use-CallerPreference -Cmdlet $PSCmdlet -SessionState $ExecutionContext.SessionState
 
-    $nuGetPath = Install-WhiskeyNuGet -DownloadRoot $BuildRoot -Version $Version
+    $NuGetPath = Join-Path -Path $NuGetPath -ChildPath 'tools\NuGet.exe' -Resolve
+    if( -not $NuGetPath )
+    {
+        Stop-WhiskeyTask -Context $TaskContext -Message "NuGet.exe not found at ""$($NuGetPath)""."
+        return
+    }
 
     foreach( $item in $Path )
     {