+++ CHANGELOG.md
@@ -1,7 +1,28 @@
 
-# 0.49.3
+# 0.50.0
+
+## Upgrade Instructions
+
+Whiskey now requires PackageManagement 1.4.7 and PowerShellGet 2.2.5 to be installed on any computer on which it runs.
+Please ensure these modules are installed.
+
+## Added
+
+* Whiskey's "Version" task now calculates the next prerelease version number for a package by querying the package's
+  package repository.
+
+## Changed
+
+* Whiskey now depends on and requires PackageManagement 1.4.7 and PowerShellGet 2.2.5. These must be pre-installed on
+computers running whiskey.
+
+## Fixed
 
 * Fixed: Parallel task fails if it runs a custom task that was imported from a module.
+
+## Removed
+
+* Removed OpenCover support from the NUnit2 and NUnit3 tasks.
 
 
 # 0.49.2